[PLAT-15898][PLAT-15900][PLAT-15960][PLAT-15901]: Add support for PDB policy on k8s universes.
Summary:
Added the changes to create and delete a PDB on namespace scope for both primary and read-replica clusters during the create universe task.
Also, added the support to create PDB on the old universe during the software upgrade having the new Helm naming style.
Older PDB through helm charts will be skipped as we create PDB on the namespace scope.
PDB will only be created on the universe using the new helm naming style.
PDB will work correctly only on the universe with a single namespace/cluster. Still, in the case of multi-namespace or cluster, we will deploy the same PDB with maxUnavailable as floor(rf/2) to each namespace/cluster as our best effort.

Test Plan:
Tested manually by creating the multi-az and multi-namespace universe and verified that PDB is created for master/tserver in the primary cluster and tserver for read-replica cluster.
Verified that match labels in PDB cover all the required pods.
Tested that PDB policies are generated on the old universe during software upgrade having new Helm naming style.

Reviewers: #yba-api-review, vkumar, sneelakantan, anijhawan

Reviewed By: #yba-api-review, anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39864