[BACKPORT 2.20][#11117] DocDB: Set a default max size for RocksDB Manifest and make it configurable via Flags
Summary:
The MANIFEST file is a combination of:

  - Initial set of SST files for the RocksDB (since the last process restart)
  - A log of all SST files added and removed from the RocksDB instance since the last restart.

Default max size of the MANIFEST is UINT64MAX, so the MANIFEST file never rolls over until there is a restart of the process when the rollover logic kicks in.

This can cause the size to grow to the order of GB consuming significant diskbandwidth during copies of the file - which happens during PITR snapshots, RBS etc.

Configuring a default max size of 10MB which should cause the MANIFEST size to automatically rollover once it reaches the limit. 10MB is still sufficient to maintain a large number of tables per tablet.

Making it a Flag so that it can be configurable dynamically if needed.
Jira: DB-1026

Original commit: 8d767d76308ff06f5920d49c4ab1db53d781b776 / D39660

Test Plan:
ybt db_test DBTest.ManifestRollOver
ybt yb-admin-snapshot-schedule-test YbAdminSnapshotScheduleTest.Basic

Manually caused the file to grow to several GB and triggered a rollover as well.

Reviewers: timur, jhe

Reviewed By: jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39867