[#24855] DocDB: Refactor PgsqlReadOperation to minimize number of duplicate argument passing
Summary:
PgsqlReadOperation accepts a lot of arguments in Execute method, then passing most of them to underlying methods.
Instead of that we could pass those arguments in ctor and store it in the class.
So would not have to pass them in multiple places.

Also to minimise list of passed arguments, could introduce struct to pass common arguments in HandlePgsqlReadOperation.

Also moved GetIntents out of PgsqlReadOperation.
Jira: DB-13969

Test Plan: Jenkins

Reviewers: arybochkin, aleksandr.ponomarenko, timur

Reviewed By: arybochkin

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D39854