[#24641] YSQL: Fix planner updated col list computation for generated cols
Summary:
Postgres 12 introduced support for generated columns. This revision targets UPDATE queries that impact generated columns.

From Postgres documentation:
> Generated columns maintain access privileges separately from their underlying base columns. So, it is possible to arrange it so that a particular role can read from a generated column but not from the underlying base columns.

In Yugabyte, the planner uses `rte->updatedCols` as a proxy for identifying what columns of a given relation are being updated by the query. This `updatedCols` is a set of columns that need UPDATE permissions for the given query and are computed at the parsing stage from the SET target list of the query. Since, generated columns are NOT allowed to be a part of the SET target list, `updatedCols` always excludes generated columns that would be updated by the query due to a change in values of any of their respective parent columns.
In Yugabyte, this set of modified columns is used to:
 - Compute what indexes and constraint checks can be skipped if their associated columns are not modified
 - Determine if and when to use the scanned tuple from the child node of a ModifyTable

Since updatedCols excludes generated columns, this can cause problems in query execution further downstream.
The most notable issue is that it can cause index inconsistencies due to missed updates on indexes composed entirely of generated columns and unmodified columns.

This revision replaces all usages of `updatedCols` in the planner with a set of columns that includes both:
 - Columns specifically having UPDATE privileges
 - Any generated columns that are dependent on them
Jira: DB-13705

Test Plan:
```
./yb_build.sh debug --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflict#schedule'
```

Reviewers: amartsinchyk, aagrawal

Reviewed By: amartsinchyk, aagrawal

Subscribers: smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D39519