[#23493] xCluster: Make bootstrapping of sequences_data write out current table into the WAL on the source universe side
Summary:
As title.  As part of inbound replication group setup, we now make an
RPC to the source universe master, EnsureSequenceUpdatesInWalRPC, which
writes out the current sequence information to the WAL.

Note to reviewers: please check the new RPC code carefully; I don't
entirely understand the framework.   I did not add code for retrying in
some cases that I saw in nearby RPC code.

Upgrade/Downgrade safety:
In terms of upgrades safety, this new code only runs in automatic mode when the test flag for sequence replication is on.    This will be guarded by an auto flag eventually.

Fixes #23493

Test Plan:
Added several end to end tests to make sure that bootstrapping of sequence
information works even when we are bumping sequences during the
bootstrapping.

Also added a test to make sure that this RPC failing does not crash the
server.  (An earlier version did this because calling abort on the
reactor thread crashes the process.)

```
ybd --cxx-test xcluster_sequences-test --gtest_filter '*.SequenceReplication*'
```

Reviewers: xCluster, hsunder

Reviewed By: hsunder

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39973