[#24890] YSQL: Temporarily disable PgsqlTestMigrationFromEarliestSysCatalogSnapshot
Summary:
Currently, PgsqlTestMigrationFromEarliestSysCatalogSnapshot tries to restore a pg11 snapshot
on a pg15 cluster, which is not supported. This test fails with a weird error message
(https://github.com/yugabyte/yugabyte-db/issues/24718 tracks the fix for this).

Temporarily disable YbAdminSnapshotScheduleUpgradeTestWithYsqlPg11.PgsqlTestMigrationFromEarliestSysCatalogSnapshot
until https://github.com/yugabyte/yugabyte-db/issues/24718 is implemented. Then we can re-enable
this test and use it to check that the right error message is being thrown.
Jira: DB-14012

Test Plan:
```
./yb_build.sh --cxx-test yb-admin-snapshot-schedule-test
```

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39945