[#24916] YSQL: Fix DocDB result rows estimate in partial index scan
Summary:
Before this change, the cost of a partial index was being over-estimated. This
was caused because we did not include the partial index predicate in estimating
the number of rows returned from DocDB. This resulted in over-estimation of
network costs.

This bug has been fixed in this change.
Jira: DB-14047

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPlanner'

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D39999