[#24761] YSQL: Persist ysql major catalog upgrade state and remove test flag
Summary:
Store the Ysql Major upgrade catalog upgrade state on disk and use that to drive the tasks and checks. This eliminates the need for `FLAGS_TEST_online_pg11_to_pg15_upgrade`, and prevents invalid transitions.
The states are:
    INVALID
    DONE
    FAILED
    PERFORMING_INIT_DB
    PERFORMING_PG_UPGRADE
    MONITORING
    PERFORMING_ROLLBACK

The sequence of allowed state transitions is captured in `kAllowedTransitions` of `ysql_catalog_config.cc`.

The current ysql major catalog version number is now stored in `YsqlMajorUpgradeInfoPB`. This value is retrieved from the `VersionInfo`. `StartYsqlMajorVersionUpgrade` RPCs will fail is we are already on the latest ysql major version.

- Replaced `YBIsTestOnlinePg11ToPg15Upgrade` with `YBIsMajorUpgradeInitDb` and env var `YB_PG_MAJOR_UPGRADE_INITDB`. This is set only during ysql major catalog upgrades.
- On yb-master restart if we were in the middle of initdb, pg_upgrade or rollback transition to FAILED state.
- Replaced `test_upgrade_simultaneous_rpcs.sh`, `test_upgrade_prohibited_rpcs.sh`, `test_upgrade_rollback.sh` and `test_upgrade_double_initdb.sh  with `ysql_major_upgrade_rpcs-test.cc` and added yb-admin tests.

**Upgrade/Rollback safety:**
Modified proto are not available in any existing released builds. So it is safe to change these and not worry about rollbacks.
On an upgrade of an older universe `catalog_version` will default to version 11 and the state will default to `DONE`.
When a new cluster is created on a build with this diff, the `catalog_version` will explicitly be set to 15.

Fixes #24761
Jira: DB-13851

Test Plan:
ysql_major_upgrade_rpcs-test

pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh
```
0	2024-11-13T10:08:25-08:00	SHELL	test_upgrade_preflight
0	2024-11-13T10:09:49-08:00	SHELL	test_upgrade_index
0	2024-11-13T10:12:31-08:00	SHELL	test_upgrade_rollback
0	2024-11-13T10:14:03-08:00	SHELL	test_upgrade
0	2024-11-13T10:15:13-08:00	SHELL	test_upgrade_dropped_col
0	2024-11-13T10:16:28-08:00	SHELL	test_upgrade_globals
```

Reviewers: fizaa, tfoucher, telgersma

Reviewed By: tfoucher, telgersma

Subscribers: yql, yugaware, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39680