[#24947] DocDB: Fetch intents DB when querying vector index
Summary:
The vector index is updated only when transaction is applied.
So we could get into situation when transaction was committed, but not yet applied.
In this case we should scan intents DB for updated vectors and take them into account when populating read result.
Jira: DB-14085

Test Plan: PgVectorIndexTest.NotApplied/*

Reviewers: arybochkin

Reviewed By: arybochkin

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40042