[#25014] YSQL: change some test functions to return Status
Summary:
If a C++ test calls a function which has an ASSERT, the test does not
halt at that point, particularly for multithreaded tests.  It is better
to have these functions return Status or Result so that the test itself
can do the ASSERTs.

Make that transformation for the BumpCatalogVersion function and
functions related to that.
Jira: DB-14155

Test Plan:
Close: #25014
Jenkins: compile only

Reviewers: myang, dmitry

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D40110