[PLAT-15921] Fix resuming xcluster universe where taskParams populated incorrectly
Summary: Fix issue where the universeDetails set correctly during the `ResumeXClusterUniverses` task. During different task creation methods, we need to set the taskParams for the correct universe. This should fix the issue with YBM's case where `rootAndClientRootCASame` in universe details was not set correctly, causing the `ResumeXClusterUniverses` task to fail.

Test Plan:
Create 2 universes with the source universe with `rootAndClientRootCASame` as false and `rootAndClientRootCASame` for the target universe as true.

Set up DR between these two universes and then perform `PauseXClusterUniverses` and then `ResumeXClusterUniverses`. Make sure that on the `ResumeXClusterUniverses` task, check that the correct parameters are populated (`rootAndClientRootCASame` is set as false for source universe subtasks while it is true for target universe subtasks).

Reviewers: hzare, amindrov, spothuraju

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40113