[PLAT-13338]Redacting cert keys from helm value outputs
Summary:
We are using helm upgrade with --debug which causes prints the all the user applied values. If universe has TLS, then this will log the entire private key.
We are trying to grep any line that has //key:// and then checking if the value of that key is base64 encoded, if yes then we are redacting it.

Test Plan:
dev-itests
tested locally
{F312104}

{F312103}

Reviewers: vkumar, anijhawan, svarshney, asharma, amalyshev

Reviewed By: anijhawan, amalyshev

Subscribers: amalyshev, svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40177