[#24824] DocDB: Fix bound for backward scan
Summary:
When user does not specify bounds for backward scan, we use -Inf as a bound.
But -Inf is encoded as kLowest (0).
There are could be special rows with key greater than kLowest, for instance Transaction Apply State.
So we could use kNullLow as bound in case of backward scan with -Inf bound.
Jira: DB-13928

Test Plan: PgSingleTServerTest.BoundedBackwardScanWithLargeTransaction

Reviewers: arybochkin, rthallam

Reviewed By: arybochkin

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40015