[PLAT-16082]: UI Incorrectly Allows Rolling Restart for K8s Operator-Created Universes
Summary:
he UI currently allows users to initiate a rolling restart for universes created and managed by the Kubernetes (K8s) operator. However, this operation is restricted by the API, and any attempt to execute it results in a failure with a "403 Forbidden" error. The UI
should not enable the rolling restart option for such universes.
This diff disabled the rolling restart action when the universe is k8 controlled

Test Plan: Tested locally

Reviewers: daniel, anijhawan

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40144