[#25016] YSQL: Update upgrade-from for 2.25 to a recent commit
Summary:
The version we use as the upgrade-from for 2.25 in the tests is old. One problem is that it doesn't set ysql_major_catalog_upgrade_info.catalog_version properly to 15 on a clean install (or perhaps it loses the setting in a protobuffer copy from the sys catalog snapshot). Update the 2.25 version to recent commit 6ad0081e6a0d760a657ae5cf43f376086b15edce AKA 2.25.0.0-b340.
Jira: DB-14157

Test Plan:
Jenkins: test regex: Upgrade

On MacOS 14 arm64:
  ./yb_build.sh release --cxx-test integration-tests_basic_upgrade-test --gtest_filter UpgradeFrom_2_25_0_0/BasicUpgradeTest.TestUpgrade/0 --with_tests
  ./yb_build.sh fastdebug --cxx-test integration-tests_basic_upgrade-test --gtest_filter UpgradeFrom_2_25_0_0/BasicUpgradeTest.TestUpgrade/0 --with_tests
(to test builds.xml builds for Mac debug, primarily, because that build flavor isn't exercised by the automated tests right now)

Also, in private code I added logging that the catalog_version field is now set correctly for a clean install in the upgrade-to for the above test.

Reviewers: hsunder

Reviewed By: hsunder

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40159