[PLAT-16094] Show restart replication banner only for missing op id replication error
Summary:
The restart replication banner shows whenever there is a table in error status.
This was correct when we assigned error status only when replication was broken and a restart is required.
But since we now assign the error status to mismatched schema tables, we need to add a check for replication
error and ensure we're showing the restart replication banner only when replication error is `Missing op id`

Test Plan:
Set up xCluster DR config.
Add a column to a table under replication from the source universe.
Verify the table is marked with `Schema mismatch` and no replication restart banner is shown.

Break replication.
Verify that the replication restart banner eventually shows up.
{F312763}

Reviewers: hzare, rmadhavan, cwang

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40160