[#24106][#24107] CDC: Support lsn_type in CREATE_REPLICATION_SLOT ddl
Summary:
We need to provide the ability to compare two LSNs regardless of the slots they belong to. This capability can be utilised in various scenarios, enhancing the flexibility of taking the snapshot of the tables, streaming its data in different sessions.

To support the above requirement, the following new syntax will be introduced and supported, with the addition of the options to specify `lsn_type` as `HYBRID_TIME` or `SEQUENCE`:

```
CREATE_REPLICATION_SLOT slot_name [ TEMPORARY ][HYBRID_TIME|SEQUENCE] [USE_SNAPSHOT|EXPORT_SNAPSHOT|]
```

* The optional value `[HYBRID_TIME|SEQUENCE]` denotes the LSN type. The default value will be `SEQUENCE` and it will be valid in the context of a slot. This is a monotonic increasing number that will determine the record in global order within the context of a slot. However, this LSN can’t be compared across two LSN’s of different slots.
* The value `HYBRID_TIME` will mean that the LSN will be denoted by the `HybridTime` of the transaction commit record. All the records of the transaction that is streamed will have the same LSN as that of the commit record. The user has to ensure that the changes of a transaction are applied in totality and the acknowledgement is sent only if the commit record of a transaction is processed. With this mode, the LSN value can be compared across the different slots.

**Upgrade / rollback safety:**

This diff adds a new field for `lsn_type` to the following proto files:
1. `pg_client.proto`
2. `common.proto`
3. `master_replication.proto`

A new `kLocalPersisted` flag is introduced i.e. `ysql_yb_allow_replication_slot_lsn_types` and the value of this flag will be checked whenever the fields for `lsn_type` will be modified.
Jira: DB-12995, DB-12996

Test Plan:
The diff introduced some tests and they can be run using:

```
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestPgCreateReplicationSlotDefaultLsnType
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestCreateReplicationSlotWithLsnTypeSequence
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestCreateReplicationSlotWithLsnTypeHybridTime
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestReplicationSlotLsnTypePresentAfterRestart
```

Reviewers: skumar, stiwary, xCluster, hsunder, utkarsh.munjal, sumukh.phalgaonkar

Reviewed By: hsunder, utkarsh.munjal

Subscribers: sumukh.phalgaonkar, svc_phabricator, yql, ybase, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D38331